requirements.txt
apache-airflow[s3,gcp_api]
click
overpass
requests
PyGithub
pandas
PyYaml